requirements.txt
PyGithub==2.1.1
requests==2.31.0
matplotlib==3.8.0
numpy==1.26.0

